// src/proto/post/rpc_update_post.ts
export interface UpdatePostRequest {
    /**
     * @generated from protobuf field: string post_id = 1;
     */
    postId: string;
    /**
     * @generated from protobuf field: optional string title = 2;
     */
    title?: string;
    /**
     * @generated from protobuf field: optional string short_desc = 3;
     */
    shortDesc?: string;
    /**
     * @generated from protobuf field: optional string description = 4;
     */
    description?: string;
}
/**
 * @generated from protobuf message pb.UpdatePostResponse
 */
export interface UpdatePostResponse {
    /**
     * @generated from protobuf field: pb.Post post = 1;
     */
    post?: Post;
}
// @generated message type with reflection information, may provide speed optimized methods
class UpdatePostRequest$Type extends MessageType<UpdatePostRequest> {
    constructor() {
        super("pb.UpdatePostRequest", [
            { no: 1, name: "post_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "title", kind: "scalar", opt: true, T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "short_desc", kind: "scalar", opt: true, T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "description", kind: "scalar", opt: true, T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<UpdatePostRequest>): UpdatePostRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.postId = "";
        if (value !== undefined)
            reflectionMergePartial<UpdatePostRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: UpdatePostRequest): UpdatePostRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string post_id */ 1:
                    message.postId = reader.string();
                    break;
                case /* optional string title */ 2:
                    message.title = reader.string();
                    break;
                case /* optional string short_desc */ 3:
                    message.shortDesc = reader.string();
                    break;
                case /* optional string description */ 4:
                    message.description = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: UpdatePostRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string post_id = 1; */
        if (message.postId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.postId);
        /* optional string title = 2; */
        if (message.title !== undefined)
            writer.tag(2, WireType.LengthDelimited).string(message.title);
        /* optional string short_desc = 3; */
        if (message.shortDesc !== undefined)
            writer.tag(3, WireType.LengthDelimited).string(message.shortDesc);
        /* optional string description = 4; */
        if (message.description !== undefined)
            writer.tag(4, WireType.LengthDelimited).string(message.description);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message pb.UpdatePostRequest
 */
export const UpdatePostRequest = new UpdatePostRequest$Type();